requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==2.0.12
comtypes==1.4.7
idna==3.10
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pypiwin32==223
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
# pyttsx3==2.90  
pywin32==306
pywin32-ctypes==0.2.3
requests==2.27.1
setuptools==75.1.0
six==1.16.0
urllib3==2.2.3
websocket-client==1.4.0
wheel==0.44.0
